+++ src/state/surveys.js
@@ -2,7 +2,6 @@ import { database } from '../firebaseConfig'
 
 const SET_SURVEYS = 'surveys/SET_SURVEYS'
 
-const myApiUrl = 'https://survey-app-84f53.firebaseio.com/surveys'
 
 export const setSurveysAction = (data) => (
     {
@@ -21,31 +20,6 @@ export const initSurveysSync = () => (dispatch, getState) => {
             })
             dispatch(setSurveysAction(firebaseData))
         })
-}
-
-export const saveNewSurvey = (newSurveyData, questions) => (dispatch, getState) => {    
-    const request = {
-        method: 'POST',
-        body: JSON.stringify(newSurveyData)
-    }
-
-    return fetch(`${myApiUrl}.json`, request)
-        .then(r => r.json())
-        .then(data => {
-            const newSurveyKey = data.name
-
-            const questionSavePromises = questions.map((question) => {
-                const request = {
-                    method: 'POST',
-                    body: JSON.stringify(question)
-                }
-
-                return fetch(`${myApiUrl}/${newSurveyKey}/questions.json`, request)
-            })
-
-            return Promise.all(questionSavePromises)
-        })
-
 }
 
 const initialState = {
