+++ src/store.js
@@ -2,12 +2,13 @@ import { createStore, combineReducers, compose, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 import surveys, { initSurveysSync } from './state/surveys'
 import snackBar from './state/snackBar'
+import newSurvey from './state/newSurvey'
 
 
 const reducer = combineReducers({
     surveys,
     snackBar,
-   
+    newSurvey,
 })
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
